Add config param for default execution mode of top-level classes
The new `junit.jupiter.execution.parallel.mode.classes.default`
configuration parameter can be used to run top-level classes in parallel
but their methods sequentially or vice versa.

Resolves #1682.
